satisfaction calculation of employee is fixed, animal not

+++ src/hva/core/Animal.java
@@ -86,13 +86,8 @@ public class Animal extends NamedEntity {
    * 
    * @return the animal's satisfaction level
    */
-  public int calculateSatisfaction() {
-    return 20 
-      + 3 * _habitat.getNumAnimalsSameSpecies(_species) 
-      - 2 *
-      (_habitat.getNumAnimals() - _habitat.getNumAnimalsSameSpecies(_species)
-      + (_habitat.area() / _habitat.getNumAnimals()) 
-      + _habitat.identifyInfluence(_species));
+  public double calculateSatisfaction() {
+    return 20 + (3 * _habitat.getNumAnimalsSameSpecies(_species)) - (2 * (_habitat.getNumAnimals() - _habitat.getNumAnimalsSameSpecies(_species))) + (_habitat.area() / _habitat.getNumAnimals()) + _habitat.identifyInfluence(_species);
   }
 
   /**
